Consolidate visibility control files. still can't use forward position controller normally
